fixed issues with supporting multiple tts engines. moved some logic to flows ready from common support
