+++ flows/customStructs.go
@@ -1,4 +1,4 @@
-package ttsChanger
+package flows
 
 import "time"
